mahakumbh-disaster-prediction/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.8.2
seaborn==0.12.2
plotly==5.16.1
geopy==2.3.0
python-dateutil==2.8.2
aiohttp==3.9.1
asyncio==3.4.3
websockets==11.0.3